requirements.txt
fastapi
uvicorn
langchain
openai
faiss-cpu
pymongo
python-dotenv
tiktoken
